# Demo/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "My user project Application",
        "license": {
            "name": "MIT",
            "url":"https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:8080",
    "basePath": "/",
...
    "schemes": ["http"],
    "consumes":["application/json"],
    "produces":["application/json"],
    "paths": {
        "/utente": {
            "get": {
                "tags": ["Utente"],
                "summary": "Get all utenti in system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Utente"
                        }
                    }

                }

            },
            "post": {
                "tags": [
                    "Utente"
                ],
                "summary": "Post utenti in system",
                "parameters": [
                        {
                        "in":"body",
                        "name":"body",
                        "description":"Api for users in the system",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/Utente"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Utente"
                        }
                    }
                }
            }
            
        },
        "/utente/{id}": {
            "get": {
                "tags": [
                    "Utente"
                ],
                "summary": "Get all utenti in system",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "Api for users in the system",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Utente"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Utente"
                ],
                "summary": "Delete utenti in system",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "Api for users in the system",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Utente"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Utente": {
            "required": [
                "name",
                "surname",
                "year",
                "CF",
                "email",
                "password",
                "account_type",
                "titolo_studio",
                "indirizzo",
                "SPID"

            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "year": {
                    "type": "date"
                },
                "CF": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "account_type": {
                    "type": "string"
                },
                "titolo_studio": {
                    "type": "string"
                },
                "indirizzo": {
                    "type": "string"
                 }
            }
        },
...
    }
}